requirements.txt
fastapi==0.103.2
pydantic-settings==2.6.1
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
SQLAlchemy==2.0.21
pymysql==1.1.0
alembic==1.12.0
pytest==7.4.3
httpx==0.24.1
email-validator ==2.3.0
passlib[bcrypt]==1.7.4
fastapi-mail == 1.5.0
Jinja2 == 3.1.6
PyJWT == 2.10.1
python-multipart == 0.0.20


